backend/retrieval-validation/requirements.txt
# Backend Retrieval Validation Service Requirements

# Vector database
qdrant-client==1.7.0

# Embeddings
cohere==4.3.1

# Configuration and environment
python-dotenv==1.0.0

# Testing and validation
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
tqdm==4.66.1
numpy==1.24.3

# Text processing
scikit-learn==1.3.0
nltk==3.8.1

# Logging
structlog==23.1.0

# CLI
click==8.1.7

# Data processing
pandas==2.0.3